Workaround for modern Windows versions denying background application windows from stealing focus.
Likely caused by secondary instance being the process allowed to set itself as foreground, but since CRCE is single instance the single instance attempts to set itself as foreground, which Windows doesn't like.

This only affects Normal or Maximized windows; the workaround is to set the app as Minimized prior to calling RestoreToFront().